+++ weblate/accounts/templatetags/authnames.py
@@ -120,3 +120,14 @@ def key_name(device: Device) -> str:
 @register.simple_tag
 def second_factor_name(name: DeviceType) -> str:
     return SECOND_FACTORS[name]
+
+
+@register.simple_tag(takes_context=True)
+def format_site_title(context) -> str:
+    style = ""
+    site_title = context["site_title"]
+    if context["support_status"]["is_hosted_weblate"]:
+        style = "text-info"
+        site_title = "Weblate cloud"
+
+    return format_html('<span class="{}">{}</span>', style, site_title)