+++ weblate/addons/forms.py
@@ -518,7 +518,9 @@ class CDNJSForm(BaseAddonForm):
         max_value=100,
         min_value=0,
         required=True,
-        help_text=gettext_lazy("Threshold for inclusion of translations."),
+        help_text=gettext_lazy(
+            "The percentage of translated strings that must be present for translation to be included."
+        ),
     )
     css_selector = forms.CharField(
         label=gettext_lazy("CSS selector"),
@@ -590,31 +592,31 @@ class PseudolocaleAddonForm(BaseAddonForm):
     )
     # This shadows prefix from the Form class
     prefix = forms.CharField(  # type: ignore[assignment]
-        label=gettext_lazy("Fixed string prefix"),
+        label=gettext_lazy("Prepended static text"),
         required=False,
         initial="",
     )
     var_prefix = forms.CharField(
-        label=gettext_lazy("Variable string prefix"),
+        label=gettext_lazy("Prepended variable text"),
         required=False,
         initial="",
     )
     suffix = forms.CharField(
-        label=gettext_lazy("Fixed string suffix"),
+        label=gettext_lazy("Appended static text"),
         required=False,
         initial="",
     )
     var_suffix = forms.CharField(
-        label=gettext_lazy("Variable string suffix"),
+        label=gettext_lazy("Appended variable text"),
         required=False,
         initial="",
     )
     var_multiplier = forms.FloatField(
-        label=gettext_lazy("Variable part multiplier"),
+        label=gettext_lazy("Variable text multiplier"),
         required=False,
         initial=0.1,
         help_text=gettext_lazy(
-            "How many times to repeat the variable part depending on "
+            "How many times to repeat the variable text depending on "
             "the length of the source string."
         ),
     )
@@ -679,19 +681,27 @@ class ChangeBaseAddonForm(BaseAddonForm):
     )
 
 
-class WebhooksAddonForm(ChangeBaseAddonForm):
+class BaseWebhooksAddonForm(ChangeBaseAddonForm):
     """Form for webhook add-on configuration."""
 
     webhook_url = WeblateServiceURLField(
         label=gettext_lazy("Webhook URL"),
         required=True,
     )
+
+    field_order = ["webhook_url", "events"]
+
+
+class WebhooksAddonForm(BaseWebhooksAddonForm):
+    """Form for webhook add-on configuration."""
+
     secret = forms.CharField(
         label=gettext_lazy("Secret"),
         validators=[
             validate_base64_encoded_string,
         ],
         required=False,
+        help_text=gettext_lazy("A Base64 encoded string"),
     )
 
     field_order = ["webhook_url", "secret", "events"]