+++ weblate/formats/models.py
@@ -94,6 +94,7 @@ class FormatsConf(AppConf):
         "weblate.formats.ttkit.RESXFormat",
         "weblate.formats.ttkit.AndroidFormat",
         "weblate.formats.ttkit.MOKOFormat",
+        "weblate.formats.ttkit.CMPFormat",
         "weblate.formats.ttkit.JSONFormat",
         "weblate.formats.ttkit.JSONNestedFormat",
         "weblate.formats.ttkit.WebExtensionJSONFormat",